util: move git_online_cpus into util
The number of CPUs is useful information for creating a thread pool or a
number of workers, but it's not really about threading directly.  Evict
it from the thread file
